+++ Security/OAuthUtils.php
@@ -90,16 +90,34 @@ class OAuthUtils
      */
     public function getAuthorizationUrl(Request $request, $name, $redirectUrl = null, array $extraParameters = array())
     {
+        $resourceOwner = $this->getResourceOwner($name);
         if (null === $redirectUrl) {
             if (!$this->connect || !$this->securityContext->isGranted('IS_AUTHENTICATED_REMEMBERED')) {
                 $redirectUrl = $this->httpUtils->generateUri($request, $this->ownerMap->getResourceOwnerCheckPath($name));
             } else {
-                $request->attributes->set('service', $name);
-                $redirectUrl = $this->httpUtils->generateUri($request, 'hwi_oauth_connect_service');
+                $redirectUrl = $this->getServiceAuthUrl($request, $resourceOwner);
             }
         }
 
-        return $this->getResourceOwner($name)->getAuthorizationUrl($redirectUrl, $extraParameters);
+        return $resourceOwner->getAuthorizationUrl($redirectUrl, $extraParameters);
+    }
+
+    /**
+     * @param Request                $request
+     * @param ResourceOwnerInterface $resourceOwner
+     *
+     * @return string
+     */
+    public function getServiceAuthUrl(Request $request, ResourceOwnerInterface $resourceOwner)
+    {
+        if ($resourceOwner->getOption('auth_with_one_url')) {
+            $redirectUrl = $this->httpUtils->generateUri($request, $this->ownerMap->getResourceOwnerCheckPath($resourceOwner->getName())).'?authenticated=true';
+        } else {
+            $request->attributes->set('service', $resourceOwner->getName());
+            $redirectUrl = $this->httpUtils->generateUri($request, 'hwi_oauth_connect_service');
+        }
+
+        return $redirectUrl;
     }
 
     /**