+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 Changelog
 =========
 
+## 0.3.4 (2014-05-12)
+* Fix: Instagram OAuth redirect to one url,
+* Fix: `FOSUBUserProvider` should also implement `UserProviderInterface`,
+* Fix: `YahooResourceOwner` `infos_url` to use new format,
+* Fix: Send authorization via headers instead of URL parameter,
+* Fix: `GithubResourceOwner` revoke method,
+* Fix: Add login routing documentation note
+
 ## 0.3.3 (2014-02-17)
 * Fix: Incorrect redirect URL when no parameters are set,
 * Fix: Add missing parameter `prompt` for `GoogleResourceOwner`,